panel refaktored
