--- src/components/Admin.js
@@ -1,4 +1,0 @@
-
-
-
-
