+++ src/components/elements/CreateTicket/CreateTicket.js
@@ -1,0 +1,12 @@
+import { Box, Typography, Button } from "@mui/material";
+
+const CreateTicket = () => {
+    return (
+        <Box flexDirection='column'>
+            <Typography> 1, 2, 3, 4, 5 </Typography>
+            <Button> Megjatszom </Button>
+        </Box>
+    )
+};
+
+export default CreateTicket;
