+++ src/layout/Layout.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import {Routes, Route } from 'react-router-dom';
 import SplashPage from '../components/Splash';
-import PlayerPanel from '../components/Player';
+import UserPanel from '../components/User';
 
 const MainRoutes = () => {
     return (
         <Routes>
             <Route path='' element={<SplashPage />}/>
-            <Route path='/player' element={<PlayerPanel />} />
+            <Route path='/user' element={<UserPanel />} />
         </Routes>
     );
 };
